fix: align battery Health percentage format with Level
Health used %5d%% (integer, no decimal) while Level used %5.1f%%
(one decimal place), causing two extra spaces before the Health
percentage in the status TUI.

+++ cmd/status/view.go
@@ -752,7 +752,7 @@ func renderBatteryCard(batts []BatteryStatus, thermal ThermalStatus) cardData {
 
 		// Add capacity line if available.
 		if b.Capacity > 0 {
-			capacityText := fmt.Sprintf("%5d%%", b.Capacity)
+			capacityText := fmt.Sprintf("%5.1f%%", float64(b.Capacity))
 			if b.Capacity < 70 {
 				capacityText = dangerStyle.Render(capacityText)
 			} else if b.Capacity < 85 {
